# core/internal/transport/http/venue/types.gen.go
// Package venue provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package venue

import (
	"time"
)

const (
	ApiKeyAuthScopes = "ApiKeyAuth.Scopes"
)

// Defines values for UpdateOrderStatusRequestStatus.
const (
	UpdateOrderStatusRequestStatusCOMPLETED  UpdateOrderStatusRequestStatus = "COMPLETED"
	UpdateOrderStatusRequestStatusCOOKING    UpdateOrderStatusRequestStatus = "COOKING"
	UpdateOrderStatusRequestStatusDELIVERING UpdateOrderStatusRequestStatus = "DELIVERING"
	UpdateOrderStatusRequestStatusREADY      UpdateOrderStatusRequestStatus = "READY"
)

// Defines values for ListVenueOrdersParamsStatus.
const (
	ListVenueOrdersParamsStatusACCEPTED   ListVenueOrdersParamsStatus = "ACCEPTED"
	ListVenueOrdersParamsStatusCANCELLED  ListVenueOrdersParamsStatus = "CANCELLED"
	ListVenueOrdersParamsStatusCOMPLETED  ListVenueOrdersParamsStatus = "COMPLETED"
	ListVenueOrdersParamsStatusCOOKING    ListVenueOrdersParamsStatus = "COOKING"
	ListVenueOrdersParamsStatusCREATED    ListVenueOrdersParamsStatus = "CREATED"
	ListVenueOrdersParamsStatusDELIVERING ListVenueOrdersParamsStatus = "DELIVERING"
	ListVenueOrdersParamsStatusREADY      ListVenueOrdersParamsStatus = "READY"
)

// AvailabilityRequest defines model for AvailabilityRequest.
type AvailabilityRequest struct {
	IsAvailable *bool `json:"isAvailable,omitempty"`
}

// CreateMenuItemRequest defines model for CreateMenuItemRequest.
type CreateMenuItemRequest struct {
	CategoryId   *string `json:"categoryId,omitempty"`
	Currency     *string `json:"currency,omitempty"`
	Description  *string `json:"description,omitempty"`
	Name         string  `json:"name"`
	PriceKopecks int64   `json:"priceKopecks"`
}

// Error defines model for Error.
type Error struct {
	Error *string `json:"error,omitempty"`
}

// Order defines model for Order.
type Order struct {
	CreatedAt    *time.Time `json:"createdAt,omitempty"`
	Currency     *string    `json:"currency,omitempty"`
	CustomerRef  *string    `json:"customerRef,omitempty"`
	Id           *string    `json:"id,omitempty"`
	Status       *string    `json:"status,omitempty"`
	TotalKopecks *int64     `json:"totalKopecks,omitempty"`
	VenueId      *string    `json:"venueId,omitempty"`
}

// RejectOrderRequest defines model for RejectOrderRequest.
type RejectOrderRequest struct {
	Reason *string `json:"reason,omitempty"`
}

// UpdateOrderStatusRequest defines model for UpdateOrderStatusRequest.
type UpdateOrderStatusRequest struct {
	Status *UpdateOrderStatusRequestStatus `json:"status,omitempty"`
}

// UpdateOrderStatusRequestStatus defines model for UpdateOrderStatusRequest.Status.
type UpdateOrderStatusRequestStatus string

// BadRequest defines model for BadRequest.
type BadRequest = Error

// NotFound defines model for NotFound.
type NotFound = Error

// ListVenueOrdersParams defines parameters for ListVenueOrders.
type ListVenueOrdersParams struct {
	Status *ListVenueOrdersParamsStatus `form:"status,omitempty" json:"status,omitempty"`
}

// ListVenueOrdersParamsStatus defines parameters for ListVenueOrders.
type ListVenueOrdersParamsStatus string

// CreateMenuItemJSONRequestBody defines body for CreateMenuItem for application/json ContentType.
type CreateMenuItemJSONRequestBody = CreateMenuItemRequest

// SetMenuItemAvailabilityJSONRequestBody defines body for SetMenuItemAvailability for application/json ContentType.
type SetMenuItemAvailabilityJSONRequestBody = AvailabilityRequest

// RejectOrderJSONRequestBody defines body for RejectOrder for application/json ContentType.
type RejectOrderJSONRequestBody = RejectOrderRequest

// UpdateOrderStatusJSONRequestBody defines body for UpdateOrderStatus for application/json ContentType.
type UpdateOrderStatusJSONRequestBody = UpdateOrderStatusRequest